Features
- Fix Compile Warning
- Added Free Chests and Free Shops
- RGB Sliders for menu theme (saves)
- Fix issue where it copies hacks.json to main gd directory even if it's in the GDH folder

+++ source/startposSwitcher.cpp
@@ -1,7 +1,7 @@
 #include "startposSwitcher.hpp"
 #include "smartStartpos.hpp"
 
-namespace startposSwitcher 
+namespace startposSwitcher
 {
     void(__thiscall *resetLevel)(void *) = nullptr;
     void(__thiscall *setStartPosObject)(void *, void *) = nullptr;
@@ -12,18 +12,18 @@ namespace startposSwitcher
     void *playLayer = nullptr;
     std::vector<float *> startposObjects;
     int startposIndex = -1;
-    float tmp[] = { 0, 0 };
+    float tmp[] = {0, 0};
     bool enabled = false;
     bool alternateKeys = false;
 
-    void __fastcall startposObjectInitHook(void *self, void *_) 
+    void __fastcall startposObjectInitHook(void *self, void *_)
     {
         startposObjects.push_back((float *)self);
         startposIndex = startposObjects.size() - 1;
         startposObjectInit(self);
     }
 
-    void __fastcall playLayerInitHook(void *self, void *_) 
+    void __fastcall playLayerInitHook(void *self, void *_)
     {
         startposObjects = {};
         startposIndex = 0;
@@ -32,7 +32,7 @@ namespace startposSwitcher
         playLayerInit(self);
     }
 
-    void __fastcall playLayerDestructorHook(void *self, void *_) 
+    void __fastcall playLayerDestructorHook(void *self, void *_)
     {
         playLayer = nullptr;
         playLayerDestructor(self);
@@ -51,7 +51,8 @@ namespace startposSwitcher
 
     void handleKeyPress(int keyCode)
     {
-        if (!enabled) return;
+        if (!enabled)
+            return;
         if (!alternateKeys && (keyCode == VK_LEFT || keyCode == VK_RIGHT))
         {
             switchStartPos(keyCode == VK_RIGHT);
@@ -64,23 +65,23 @@ namespace startposSwitcher
 
     void switchStartPos(bool direction)
     {
-        if (playLayer == nullptr || !enabled) 
+        if (playLayer == nullptr || !enabled)
         {
             return;
         }
 
-        if (direction) 
+        if (direction)
         {
             startposIndex = startposIndex + 1;
-            if (startposIndex >= startposObjects.size()) 
+            if (static_cast<unsigned int>(startposIndex) >= startposObjects.size())
             {
                 startposIndex = -1;
             }
         }
-        else 
+        else
         {
             startposIndex = startposIndex - 1;
-            if (startposIndex < -1) 
+            if (startposIndex < -1)
             {
                 startposIndex = startposObjects.size() - 1;
             }
@@ -88,11 +89,11 @@ namespace startposSwitcher
 
         ((uint32_t *)playLayer)[0xB7D] = 0; // PlayLayer::startPosCheckpoint i believe?
 
-        if (startposIndex >= 0) 
+        if (startposIndex >= 0)
         {
             setStartPosObject(playLayer, startposObjects[startposIndex]);
-        } 
-        else 
+        }
+        else
         {
             setStartPosObject(playLayer, nullptr);
         }
@@ -105,7 +106,7 @@ namespace startposSwitcher
         alternateKeys = alternate;
     }
 
-    void setEnabled(bool enable) 
+    void setEnabled(bool enable)
     {
         enabled = enable;
     }
